+++ crates/ruma-events/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Improvements:
 
+- Add support for the `m.room_key_bundle` to-device event, which was introduced in [MSC4268](https://github.com/matrix-org/matrix-spec-proposals/pull/4268).
 - Add support for the `m.room_key.withheld` to-device event, which was introduced in Matrix 1.1.
 - Fix and stabilize support for rich text in room topics, according to Matrix 1.15.